Add files via upload
Reorganized a few things. File writing functions are in their own file now
