+++ score.py
@@ -1,6 +1,5 @@
 
 #calculate the scores from the document 
-
 def finalscore(wins, losses, ties):
     print()
     final = int(wins)*2 + int(losses)*-1 + int(ties)
